kernel/pyproject.toml
[project]
name = "arvai-kernel"
version = "0.1.0"
description = "Arvai knowledge management service kernel"
readme = "README.md"
authors = [
    { name = "ke757", email = "liuyuxiaokk@gmail.com" }
]
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "pyyaml>=6.0",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "sqlmodel>=0.0.22",
    "aiosqlite>=0.20.0",
]

[project.scripts]
arvai-kernel = "app.main:run"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]
